Fix a warning in Core::sleep()
Thanks @jleclanche

+++ src/core/core.cpp
@@ -100,7 +100,7 @@ Core::~Core()
     conf->killInstance();
 }
 
-void Core::sleep(quint8 msec)
+void Core::sleep(int msec)
 {
     QMutex mutex;
     mutex.lock();
